backend/hackoverflow/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.2.1
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-rest-knox==4.1.0
djangorestframework==3.12.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.1
progress==1.5
pycparser==2.20
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
urllib3==1.25.8
webencodings==0.5.1
